requirements.txt
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.28.0
flask>=2.3.0
flask-cors>=4.0.0
pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
opencv-python>=4.8.0
gtts>=2.3.0
pyttsx3>=2.90
requests>=2.31.0
scikit-learn>=1.3.0
scipy>=1.11.0
seaborn>=0.12.0
plotly>=5.17.0
reportlab>=4.0.0
firebase-admin>=6.2.0
streamlit-authenticator>=0.2.3
pygame>=2.1.0